'unsigned int' replaced by size_t

+++ ql/Instruments/capfloor.hpp
@@ -60,7 +60,7 @@ namespace QuantLib {
             std::vector<Rate> exerciseRates_;
             RelinkableHandle<TermStructure> termStructure_;
             Handle<InterestRateModelling::Model> model_;
-            unsigned int nPeriods_;
+            size_t nPeriods_;
             std::vector<Time> startTimes_;
             std::vector<Time> endTimes_;
             std::vector<Time> tenors_;
