+++ ql/InterestRateModelling/CalibrationHelpers/cap.hpp
@@ -59,7 +59,7 @@ namespace QuantLib {
                 RelinkableHandle<TermStructure> termStructure_;
                 Handle<Instruments::SimpleSwap> swap_;
                 Handle<Instruments::EuropeanCap> cap_;
-                unsigned int nbOfPeriods_;
+                size_t nbOfPeriods_;
                 std::vector<Time> startTimes_;
                 std::vector<Time> endTimes_;
             };
